+++ frontend/src/pages/workspaces/[workspaceSlug]/pipelines/index.tsx
@@ -50,7 +50,7 @@ const WorkspacePipelinesPage: NextPageWithLayout = ({
         helpLinks={[
           {
             label: t("About pipelines"),
-            href: "https://github.com/BLSQ/openhexa/wiki/User-manual#using-pipelines",
+            href: "https://docs.openhexa.com/pipelines/",
           },
           {
             label: t("Writing OpenHEXA pipelines"),