+++ frontend/src/workspaces/features/DatabaseVariablesSection/DatabaseVariablesSection.tsx
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import { DatabaseVariablesSection_WorkspaceFragment } from "./DatabaseVariablesSection.generated";
+import { DatabaseVariablesSection_CredentialsFragment } from "./DatabaseVariablesSection.generated";
 import { useTranslation } from "next-i18next";
 import { useMemo } from "react";
 import DataGrid, { BaseColumn } from "core/components/DataGrid";
@@ -9,14 +9,14 @@ import Clipboard from "core/components/Clipboard";
 import { slugify } from "workspaces/helpers/connections/utils";
 
 type DatabaseVariablesSectionProps = {
-  workspace: DatabaseVariablesSection_WorkspaceFragment;
+  credentials: DatabaseVariablesSection_CredentialsFragment | null | undefined;
 };
 
-const DatabaseVariablesSection = (props: DatabaseVariablesSectionProps) => {
+const DatabaseVariablesSection = ({
+  credentials,
+}: DatabaseVariablesSectionProps) => {
   const { t } = useTranslation();
-  const {
-    database: { credentials },
-  } = props.workspace;
+
   const dbCredentials = useMemo(
     () => [
       {
@@ -54,16 +54,17 @@ const DatabaseVariablesSection = (props: DatabaseVariablesSectionProps) => {
   );
   return (
     <DataGrid
-      className="rounded-md border 2xl:w-3/4"
+      className="rounded-md border"
       data={dbCredentials}
-      fixedLayout={false}
+      fixedLayout={true}
     >
       <TextColumn
         className="py-3 font-mono"
         label={t("Name")}
         accessor="name"
+        width={100}
       />
-      <BaseColumn label={t("Environment variable")} accessor="name">
+      <BaseColumn label={t("Environment variable")} accessor="name" width={150}>
         {(value) => (
           <Clipboard value={slugify(`WORKSPACE_DATABASE_${value}`)}>
             <code className="rounded-md bg-slate-100 p-1.5 font-mono text-xs font-medium text-gray-600">
@@ -75,6 +76,7 @@ const DatabaseVariablesSection = (props: DatabaseVariablesSectionProps) => {
       <BaseColumn
         className="flex gap-x-1.5 font-mono text-gray-900"
         label={t("Value")}
+        width={400}
       >
         {(field) => (
           <div className="flex gap-x-1">
@@ -91,20 +93,15 @@ const DatabaseVariablesSection = (props: DatabaseVariablesSectionProps) => {
   );
 };
 
-DatabaseVariablesSection.fragment = {
-  workspace: gql`
-    fragment DatabaseVariablesSection_workspace on Workspace {
-      slug
-      database {
-        credentials {
-          dbName
-          username
-          password
-          host
-          port
-          url
-        }
-      }
+DatabaseVariablesSection.fragments = {
+  credentials: gql`
+    fragment DatabaseVariablesSection_credentials on DatabaseCredentials {
+      dbName
+      username
+      password
+      host
+      port
+      url
     }
   `,
 };